Finished Encoder/Decoder v0.0.0.5
+ Added comments
+ Added examples
~ Refactored some functions/data
+ Added RLE sequence
+ Added proper encoded stream write and read
+ Variefied with valgrind for mamory leaks (none)
+ Added linux build (with Win10 bash)

+++ MatrixReader.hpp
@@ -9,10 +9,13 @@
 
 
 namespace dc {
+    /**
+     *  A class to read a matrix of size <size>*<size> from a text file.
+     */
     template<size_t size = dc::BlockSize>
     class MatrixReader {
         private:
-            uint32_t matrix  [size * size];
+            uint16_t matrix  [size * size];
             double   expanded[size * size];
             std::string m_errStr;
